refactor: split cloud sandbox io primitives
